[FIX] Dossier 'Projects' en trop
