+++ src/services/api/index.ts
@@ -1,3 +1,3 @@
 export * from "./Auth";
-export * from "./Projects";
+export * from "./Project";
 export * from "./Users";
